refactor: extract row to own file, some cleanup
