+++ src/pages/organisationUnits/list/useOrganisationUnits.tsx
@@ -53,6 +53,7 @@ export const useFilteredOrgUnits = ({
         queryFn: boundQueryFn<OrganisationUnitResponse>,
         staleTime: 60000,
         cacheTime: 60000,
+        // keepPreviousData: true,
     })
 }
